backend/requirements_clean.txt
ÿþýÿýÿarrow==1.3.0



asgiref==3.9.1



binaryornot==0.4.4



certifi==2025.4.26



channels==4.3.1



channels_redis==4.3.0



chardet==5.2.0



charset-normalizer==3.4.2



click==8.2.1



colorama==0.4.6



cookiecutter==2.6.0



Django==5.2.5



django-cors-headers==4.7.0



django-environ==0.11.2



django-extensions==4.1



django-filter==25.1



django-tenants==3.9.0



djangorestframework==3.16.1



djangorestframework_simplejwt==5.5.1



idna==3.10



Jinja2==3.1.6



markdown-it-py==3.0.0



MarkupSafe==3.0.2



mdurl==0.1.2



msgpack==1.1.1



pillow==11.3.0



psycopg2-binary==2.9.10



Pygments==2.19.1



PyJWT==2.10.1



python-dateutil==2.9.0.post0



python-decouple==3.8



python-slugify==8.0.4



PyYAML==6.0.2



qrcode==8.2



redis==6.4.0



requests==2.32.3



rich==14.0.0



six==1.17.0



sqlparse==0.5.3



text-unidecode==1.3



types-python-dateutil==2.9.0.20250516



tzdata==2025.2



urllib3==2.4.0




google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.159.0
